readme updated

+++ create_jbrowse2_config.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python
 ''' parse table with tracks for jbrowse2 and create config.json file '''
+import shutil
+
 import json
 
 import subprocess
@@ -9,7 +11,7 @@ import argparse
 import os
 import sys
 import csv
-
+from shlex import quote
 
 
 
@@ -22,13 +24,17 @@ def read_config(config_file, directory):
     with open(config_file, 'r') as f:
         reader = csv.DictReader(f, delimiter="\t")
         labels = set()
+        first = True
         for record in reader:
+            if first:
+                assembly_name = record['label']
+                first = False
             if record['label'] in labels:
                 print('duplicated label:', record['label'])
                 sys.exit('labels must be unique, exiting...')
             labels.add(record['label'])
             record['new_filename'] = os.path.join(
-                directory, F"{record['label']}.{record['format']}"
+                directory, F"{assembly_name}_{record['label']}.{record['format']}"
                 )
             record['src_filename'] = os.path.join(
                 record['dirname'], record['filename']
@@ -48,11 +54,11 @@ def create_reference(ref_path, outdir, assembly_name):
     indexfile = F"{ref_path}.fai"
     if not os.path.exists(indexfile):
         # create index of fasta reference
-        cmd = F"samtools faidx {ref_path}"
+        cmd = F"samtools faidx {quote(ref_path)}"
         subprocess.check_call(cmd, shell=True)
     cmd = (
-        F"jbrowse add-assembly {ref_path} --out {outdir} --load inPlace --name "
-        F"{assembly_name} "
+        F"jbrowse add-assembly {quote(ref_path)} --out {outdir} --load inPlace --name "
+        F"{assembly_name} --overwrite"
     )
     print(cmd)
     subprocess.check_call(cmd, shell=True)
@@ -74,7 +80,7 @@ def get_config_string(track):
             "LinearWiggleDisplay",
                                 "renderers": {"XYPlotRenderer": {"color": color,
                                     "type": "XYPlotRenderer"}}}]}
-    if track['format'] == 'bed':
+    if track['format'] == 'bed' and track['type'] != 'arcs':
         config = {"displays": [{"displayId": track['label'], "type": "LinearBasicDisplay",
                                 "renderer": {"color1": color,
                                              "type": "SvgFeatureRenderer"}}]}
@@ -82,9 +88,16 @@ def get_config_string(track):
         config = {"displays": [{"displayId": track['label'], "type": "LinearBasicDisplay",
                                 "renderer": {"color1": color,
                                              "type": "SvgFeatureRenderer"}}]}
+    if track['format'] == 'bed' and track['type'] == 'arcs':
+        config = {"displays": [{"displayId": track['label'], "type": "LinearArcDisplay",
+                                "renderer": {"color": color,
+                                             "type": "ArcRenderer",
+                                             "label": ""}}]}
 
     if config!= '':
-        cfg_cmd = "--config '" + json.dumps(config)  + "'"
+        # escape quotes in config string
+        config_str = json.dumps(config).replace('"', '\\"')
+        cfg_cmd = "--config \"" + config_str  + "\""
     else:
         cfg_cmd = ''
 
@@ -102,16 +115,17 @@ def add_track(track, outdir, assembly_name):
     cfg_cmd = get_config_string(track)
     if os.path.exists(indexfile):
         # use index:
-        cmd = (F"jbrowse add-track {track['new_filename']} --out {outdir} "
-               F"--load inPlace --trackId {track['label']} --indexFile {indexfile} "
-               F" --category={track['category']}"
-               F" {cfg_cmd} assemblyNames={assembly_name} --config {config_path}")
+        cmd = (F"jbrowse add-track {quote(track['new_filename'])} --out {outdir} "
+               F"--load inPlace --trackId {quote(track['label'])} --indexFile"
+               F" {quote(indexfile)} "
+               F" --category {quote(track['category'])}"
+               F" {cfg_cmd} --assemblyNames {quote(assembly_name)} --overwrite")
         print(cmd)
     else:
-        cmd = (F"jbrowse add-track {track['new_filename']} --out {outdir} "
-               F"--load inPlace --trackId {track['label']} "
-               F" --category {track['category']}"
-               F" {cfg_cmd}")
+        cmd = (F"jbrowse add-track {quote(track['new_filename'])} --out {outdir} "
+               F"--load inPlace --trackId {quote(track['label'])} "
+               F" --category {quote(track['category'])} --overwrite"
+               F" {cfg_cmd} --assemblyNames {quote(assembly_name)}")
         print(cmd)
     try:
         subprocess.check_call(cmd, shell=True)
@@ -137,23 +151,23 @@ def sort_and_index(track):
         if os.path.exists(sortfilegz) and os.path.exists(indexfilegz):
             print(F"sorted and indexed file {sortfilegz} exists, skipping...")
         else:
-            cmd = F"sort -k1,1 -k2,2n {track['new_filename']} > {sortfile}"
+            cmd = F"sort -k1,1 -k2,2n {quote(track['new_filename'])} > {quote(sortfile)}"
             subprocess.check_call(cmd, shell=True)
-            cmd = F"bgzip {sortfile}"
+            cmd = F"bgzip {quote(sortfile)}"
             subprocess.check_call(cmd, shell=True)
-            cmd = F"tabix -C -p bed {sortfilegz}"
+            cmd = F"tabix -C -p bed {quote(sortfilegz)}"
             subprocess.check_call(cmd, shell=True)
         return sortfilegz
     # GFF
-    if track['format'] == 'gff':
+    if track['format'] == 'gff' or track['format'] == 'gff3':
         if os.path.exists(sortfilegz) and os.path.exists(indexfilegz):
             print(F"sorted and indexed file {sortfilegz} exists, skipping...")
         else:
-            cmd = F"sort -k1,1 -k4,4n {track['new_filename']} > {sortfile}"
+            cmd = F"sort -k1,1 -k4,4n {quote(track['new_filename'])} > {quote(sortfile)}"
             subprocess.check_call(cmd, shell=True)
-            cmd = F"bgzip {sortfile}"
+            cmd = F"bgzip {quote(sortfile)}"
             subprocess.check_call(cmd, shell=True)
-            cmd = F"tabix -C -p gff {sortfilegz}"
+            cmd = F"tabix -C -p gff {quote(sortfilegz)}"
             subprocess.check_call(cmd, shell=True)
         return sortfilegz
     # BAM
@@ -161,14 +175,16 @@ def sort_and_index(track):
         if os.path.exists(sortfile) and os.path.exists(indexfile):
             print(F"sorted and indexed file {sortfile} exists, skipping...")
         else:
-            cmd = F"samtools sort {track['new_filename']} -o {sortfile}"
+            cmd = F"samtools sort {quote(track['new_filename'])} -o {quote(sortfile)}"
             subprocess.check_call(cmd, shell=True)
-            cmd = F"samtools index -c {sortfile}"
+            cmd = F"samtools index -c {quote(sortfile)}"
             subprocess.check_call(cmd, shell=True)
         return sortfile
     if track['format'] == 'bigwig':
         return track['new_filename']
     print('format not recognized, skipping sorting and indexing')
+    print(track['format'])
+    print('-------------------')
     return track['new_filename']
 
 
@@ -176,7 +192,6 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('-d', '--output_dir', type=str, default="data")
     parser.add_argument("-c", "--config_table", required=True)
-    parser.add_argument("-C", "--config_path", required=False, default="config.yaml")
     parser.add_argument("-u", "--update", action="store_true", default=False)
     args = parser.parse_args()
 
@@ -186,30 +201,35 @@ if __name__ == "__main__":
     print('reference path:', ref_path)
 
     # first remove old config.json file
-    if os.path.exists(os.path.join(args.output_dir, 'config.json')):
+    if os.path.exists(os.path.join(args.output_dir, 'config.json')) and not args.update:
         print('removing old config.json file')
         os.remove(os.path.join(args.output_dir, 'config.json'))
 
     # copy all files to output directory but only if they are not already there
+    assembly_name = track_list[0]['label']
+
+
+
+
+
     for track in track_list:
         if not os.path.exists(track['new_filename']):
             print('copying file:', track['src_filename'])
             print('to:', track['new_filename'])
-            subprocess.check_call(
-                F"cp {track['src_filename']} {track['new_filename']}",
-                shell=True
-                )
+            # use shutil.copy2 to keep file metadata
+            shutil.copy(track['src_filename'], track['new_filename'])
+
         else:
             print('file already exists:', track['new_filename'])
 
     print('creating reference files:')
-    assembly_name = track_list[0]['label']
-    create_reference(ref_path, args.output_dir, assembly_name, config_path=args.config_path)
+
+    create_reference(ref_path, args.output_dir, assembly_name)
     for track in track_list[1:]:
         # sort and make index:
         track['new_filename'] = sort_and_index(track)
         print('adding track:', track['label'])
-        add_track(track, args.output_dir, assembly_name, config_path=args.config_path)
+        add_track(track, args.output_dir, assembly_name)
         print("----------------------------------------")
         print("----------------------------------------")
         print("----------------------------------------")
